refactor: restructure RoomSelectionContext following AuthContext pattern
- Move useRoomSelection hook to src/hooks/useRoomSelection.ts
- Move RoomSelectionProvider to src/components/RoomSelectionProvider.tsx
- RoomSelectionContext.tsx now only exports Context + type interface
- Remove unnecessary comment from backend/tests/people.test.ts
- Update MainView to use SidePanel instead of obsolete RoomDetails
- Delete RoomDetails.tsx component (using main's SidePanel structure)
- Add selectedPersonId feature to SidePanel and RoomOccupants
- Simplify tests: mock useRoomSelection at file level, no Provider wrapper needed
- Update all imports to use new hook and provider locations
- Fix e2e test locator scoping to dropdown
