+++ frontend/tests/MainView.test.tsx
@@ -3,7 +3,6 @@ import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it, vi } from "vitest";
 import MainView from "../src/components/MainView.tsx";
-import { RoomSelectionProvider } from "../src/contexts/RoomSelectionContext";
 import {
   DEFAULT_SCALE,
   getBottomBound,
@@ -28,12 +27,21 @@ vi.mock("../src/contexts/AuthContext", () => ({
   }),
 }));
 
-const customRender = (ui: React.ReactElement) => {
-  return render(
-    <RoomSelectionProvider fetchRoomById={vi.fn()}>
-      {ui}
-    </RoomSelectionProvider>
-  );
+vi.mock("../src/hooks/useRoomSelection", () => ({
+  useRoomSelection: () => ({
+    activeRoomId: null,
+    setActiveRoomId: vi.fn(),
+    isSidePanelOpen: false,
+    setIsSidePanelOpen: vi.fn(),
+    room: null,
+    setRoom: vi.fn(),
+    selectRoom: vi.fn(),
+    selectedPersonId: null,
+  }),
+}));
+
+const customRender = () => {
+  return render(<MainView />);
 };
 
 vi.mock("../src/assets/exactum-2.min.svg?react", () => ({
@@ -57,13 +65,13 @@ vi.mock("../src/services/roomsService");
 describe("MainView", () => {
   it("renders without crashing", () => {
     vi.mocked(findAllRooms).mockResolvedValue([]);
-    customRender(<MainView />);
+    customRender();
     expect(screen.getByTestId("mock-svg")).toBeInTheDocument();
   });
 
   it("maps rooms data correctly to SVG elements", async () => {
     vi.mocked(findAllRooms).mockResolvedValue(rooms);
-    customRender(<MainView />);
+    customRender();
 
     await waitFor(() => {
       const availableRoom = document.querySelector(
@@ -93,7 +101,7 @@ describe("MainView", () => {
 
   it("room colors change correctly when showing departments", async () => {
     vi.mocked(findAllRooms).mockResolvedValue(rooms);
-    customRender(<MainView />);
+    customRender();
 
     const user = userEvent.setup();
     user.click(screen.getByTestId("switch-color-mode"));
@@ -115,7 +123,7 @@ describe("MainView", () => {
 
   it("renders legend with correct initial mode (availability)", async () => {
     vi.mocked(findAllRooms).mockResolvedValue(rooms);
-    customRender(<MainView />);
+    customRender();
 
     await waitFor(() => {
       const legend = screen.getByTestId("legend");
@@ -125,7 +133,7 @@ describe("MainView", () => {
 
   it("legend switches to department mode when button is clicked", async () => {
     vi.mocked(findAllRooms).mockResolvedValue(rooms);
-    customRender(<MainView />);
+    customRender();
 
     const user = userEvent.setup();
 
@@ -146,7 +154,7 @@ describe("MainView", () => {
   it("fetches room details when a room is clicked", async () => {
     vi.mocked(findAllRooms).mockResolvedValue(rooms);
     vi.mocked(findRoomById).mockResolvedValue(rooms[0]);
-    customRender(<MainView />);
+    customRender();
 
     await waitFor(() => {
       const room = document.querySelector('[data-room="A210"]');
@@ -168,7 +176,7 @@ describe("MainView", () => {
 describe("MapTransform", () => {
   describe("transform with buttons", () => {
     it("zoom buttons work", async () => {
-      customRender(<MainView />);
+      customRender();
 
       const user = userEvent.setup();
       const map = document.getElementsByClassName(
@@ -191,7 +199,7 @@ describe("MapTransform", () => {
     });
 
     it("zooming with button stops at maximum", async () => {
-      customRender(<MainView />);
+      customRender();
 
       const user = userEvent.setup();
       const map = document.getElementsByClassName(
@@ -210,7 +218,7 @@ describe("MapTransform", () => {
     });
 
     it("zooming with button stops at minimum", async () => {
-      customRender(<MainView />);
+      customRender();
 
       const user = userEvent.setup();
       const map = document.getElementsByClassName(
@@ -229,7 +237,7 @@ describe("MapTransform", () => {
     });
 
     it("while zooming with buttons the map is bounded from the left and top", async () => {
-      customRender(<MainView />);
+      customRender();
 
       const user = userEvent.setup();
       const inputDiv = document.getElementsByClassName(
@@ -262,7 +270,7 @@ describe("MapTransform", () => {
     });
 
     it("while zooming with buttons the map is bounded from the right and bottom", async () => {
-      customRender(<MainView />);
+      customRender();
 
       const user = userEvent.setup();
       const inputDiv = document.getElementsByClassName(
@@ -304,7 +312,7 @@ describe("MapTransform", () => {
   describe("transform with mouse", () => {
     describe("zooming", () => {
       it("zooming in increases scale", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -325,7 +333,7 @@ describe("MapTransform", () => {
       });
 
       it("zooming out decreases scale", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -346,7 +354,7 @@ describe("MapTransform", () => {
       });
 
       it("zooming stops at maximum", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -371,7 +379,7 @@ describe("MapTransform", () => {
       });
 
       it("zooming stops at minimum", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -396,7 +404,7 @@ describe("MapTransform", () => {
       });
 
       it("zooming in moves the map outward", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -421,7 +429,7 @@ describe("MapTransform", () => {
       });
 
       it("zooming out moves the map inward", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -446,7 +454,7 @@ describe("MapTransform", () => {
       });
 
       it("while zooming the map is bounded from the left and top", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -483,7 +491,7 @@ describe("MapTransform", () => {
       });
 
       it("while zooming the map is bounded from the right and bottom", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -528,7 +536,7 @@ describe("MapTransform", () => {
 
     describe("moving", () => {
       it("moving map works", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -569,7 +577,7 @@ describe("MapTransform", () => {
       });
 
       it("moving is bounded from the left and top", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -603,7 +611,7 @@ describe("MapTransform", () => {
       });
 
       it("moving is bounded from the right and bottom", () => {
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
@@ -644,7 +652,7 @@ describe("MapTransform", () => {
 
       it("moving the map disables room hovering", async () => {
         vi.mocked(findAllRooms).mockResolvedValue(rooms);
-        customRender(<MainView />);
+        customRender();
 
         const inputDiv = document.getElementsByClassName(
           "click-container",
